Refactor directory structure
Moving profile screen components  from ui directory to components directory
