+++ community-connectivity-orlando/components/ProfileIconInside.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, TouchableOpacity, StyleSheet } from 'react-native';
-import {Colors} from "../../constants/Colors";
+import {Colors} from "../constants/Colors";
 export default function ProfileIconInside() {
     return (
             <View style={styles.container}>
